LayoutTests/networkinformation/basic-operation-expected.txt

Tests the basic operation of NetworkInfo.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS typeof connection.bandwidth is "number"
PASS typeof connection.metered is "boolean"
PASS successfullyParsed is true

TEST COMPLETE
